scraper/requirements.txt
influxdb-client
python-dotenv
paramiko
